CHANGED: HTTP GET requires LONG instead of INT now

+++ DataFlowWebservice/DataFlowWebservice/Controllers/EventsController.cs
@@ -28,7 +28,7 @@ namespace DataFlowWebservice.Controllers
         }
 
         // GET: api/Events/5
-        public ResponseModel Get(int id)
+        public ResponseModel Get(long id)
         {
             IMongoQuery query = Query<Event>.EQ(e => e.unitId, id); // Gebruikt event (e), van e check hij of het unitId en het opgegeven id hetzelfde zijn (EQ)
             return new ResponseModel(databaseManager.GetDatabase().GetCollection<Event>("events").Find(query).ToList<IResponseModel>(), ResponseModel.RESPONSE_GET);
@@ -44,7 +44,7 @@ namespace DataFlowWebservice.Controllers
         }
 
         // PUT: api/Events/5
-        public ResponseModel Put(int id, [FromBody]Event document)
+        public ResponseModel Put(long id, [FromBody]Event document)
         {
             IMongoQuery query = Query<Event>.EQ(e => e.unitId, id); // Gebruikt event (e), van e check hij of het unitId en het opgegeven id hetzelfde zijn (EQ)
             var collection = databaseManager.GetDatabase().GetCollection<BsonDocument>("events");
@@ -55,7 +55,7 @@ namespace DataFlowWebservice.Controllers
         }
 
         // DELETE: api/Events/5
-        public ResponseModel Delete(int id)
+        public ResponseModel Delete(long id)
         {
             IMongoQuery query = Query<Event>.EQ(e => e.unitId, id); // Gebruikt event (e), van e check hij of het unitId en het opgegeven id hetzelfde zijn (EQ)
             var collection = databaseManager.GetDatabase().GetCollection<BsonDocument>("events");
